Modularized MyEvents page and moved all images into image folder
