+++ friendster-frontend/vite-project/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/png" href="/friendster.png" />
+    <link rel="icon" type="image/png" href="/images/friendster.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link
   rel="stylesheet"
