+++ friendster-frontend/vite-project/src/assets/images/Friendster.png
