+++ friendster-frontend/vite-project/src/assets/images/crowd.png
