+++ friendster-frontend/vite-project/src/assets/images/event-image.jpg
