+++ friendster-frontend/vite-project/src/assets/stylesheets/_banner.scss
@@ -5,7 +5,7 @@
   justify-content: center;
   align-items: center;
   flex-direction: row;
-  background-image: url("../images/Friendster banner bg.png");
+  background-image: linear-gradient(#ffffff17, #ffffff63), url("../images/Friendster banner bg.png");
   background-size: 80%, 80%;
   background-repeat: no-repeat;
   background-position: center;
@@ -79,7 +79,7 @@
 
   .banner {
     flex-direction: column;
-    background-image: url("../images/Friendster banner_mobile.png");
+    background-image: linear-gradient(#ffffff17, #ffffff63), url("../images/Friendster banner_mobile.png");
     background-size: 65%, 65%;
   }
