+++ friendster-frontend/vite-project/src/assets/stylesheets/_footer.scss
@@ -1,6 +1,6 @@
 .footer {
   background-color: #7F5FFC;
-  background-image: linear-gradient(to bottom, #7F5FFC, #7e5ffcbd), url(../crowd.png);
+  background-image: linear-gradient(to bottom, #7F5FFC, #7e5ffcbd), url(../images/crowd.png);
   background-size: cover;
   background-position: 0% 50%;
   font-size: 20px;
